requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
frozenlist==1.3.3
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
loguru==0.7.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.3
openai==0.27.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
yarl==1.9.2
